Breakup win32.rb into more manageable chunks.
